+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.1.1] - 2023-10-06
+## [v1.1.1] - 2024-04-16
 
 ### Changed
 - Use `@shgysk8zer0/consts`
+- Update/simplify rollup config & build script
+
+### Fixed
+- Update `externals` in rollup config to use `@shgysk8zer0/consts`
 
 ## [v1.1.0] - 2023-07-03